Update config when different
Submit changes to make the remote config match the local config.

I've renamed the compare function to indicate that it's now calling updates
too. I did attempt to split out the update part, but by returning a new
`Config` item we couldn't easily log the `old -> new` value just before
changing it.

I've had to construct a new struct for the request because we need to pass a
JSON document that has a single "value" field.