Add -list-zones argument
bob asked a good question - how do you find the value for the -zone flag.

I originally got this by querying the API with curl. Then I forgot about it.
This isn't ideal for other users, so add the ability to list all zone names
and their IDs.

I've modified checkRequiredFlags() because we no longer need all of the
flags at the beginning. Depending on what route we're going to take. I've
also remove the else condition to match the early return for listZones.