Replace flag with jwaldrip/odin
This allows us to create more structured commands and reduce confusion about
which arguments you have to pass when using the zones, download or upload
functionality.

I've kept email and key as arguments, despite them being required, so that
we don't have a bewildering long list of positional arguments. In doing so
we still have to check that they don't have their default values.

`GoVersion` has changed in `Godeps.json` because I've run `godep update`
with a different version than Kush originally did. This doesn't appear to
have any significance - I can't see that it's used. However it does now
match the minimum version that we test against with Travis.