Match specific errors from CF.MakeRequest()
To ensure that we get the error that we're expecting. This highlights one
case where we're getting the wrong kind of error - a parse error instead of
"Response body indicated failure" - which I'll fix in a subsequent commit.

These are quite verbose and could be tidied up with more "public" error
objects at a later date.