Improve test for errors during CF.Update()
Part of this test was commented out because (we think) the map was being
ranged over in a different order when run on Go 1.3

Fix this by using two invalid keys (rather than one valid and one invalid)
and asserting that we should only ever see one request. That way it doesn't
matter which one is first in the range.

I've omitted the description about the public error because we're not
currently doing that and it probably belongs to Set() or MakeRequest() if
we're going to. I've instead added a description about not proceeding with
other keys, which is what we really care about.

I've also moved the handlers for valid keys into the context for testing
"logOnly true" because these should be isolated from the errors test and
"logOnly false" doesn't need any handlers as it shouldn't hit the network.